Remove Swift 3 support
