+++ CarbKit/UI/Assets.xcassets/Oval Selection.imageset/Contents.json
@@ -22,5 +22,8 @@
   "info" : {
     "version" : 1,
     "author" : "xcode"
+  },
+  "properties" : {
+    "preserves-vector-representation" : true
   }
 }
