+++ CarbKit/UI/CarbKit.storyboard
@@ -253,7 +253,7 @@
         <!--Add/Edit Carb Entry-->
         <scene sceneID="BUy-hU-2jT">
             <objects>
-                <tableViewController storyboardIdentifier="CarbEntryEditViewController" id="LyL-9U-twn" customClass="CarbEntryEditViewController" customModule="CarbKit" sceneMemberID="viewController">
+                <tableViewController storyboardIdentifier="CarbEntryEditViewController" id="LyL-9U-twn" customClass="CarbEntryEditViewController" customModule="CarbKit" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" keyboardDismissMode="onDrag" dataMode="prototypes" style="grouped" separatorStyle="default" rowHeight="44" sectionHeaderHeight="18" sectionFooterHeight="18" id="9Ci-XW-6nA">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
