+++ LoopKit/UI/TextFieldTableViewCell.swift
@@ -8,16 +8,59 @@
 
 import UIKit
 
-class TextFieldTableViewCell: UITableViewCell {
 
-    @IBOutlet var unitLabel: UILabel!
+protocol TextFieldTableViewCellDelegate: class {
+    func textFieldTableViewCellDidBeginEditing(_ cell: TextFieldTableViewCell)
+    
+    func textFieldTableViewCellDidEndEditing(_ cell: TextFieldTableViewCell)
+}
 
-    @IBOutlet var textField: UITextField!
+
+class TextFieldTableViewCell: UITableViewCell, UITextFieldDelegate {
+
+    @IBOutlet weak var unitLabel: UILabel?
+
+    @IBOutlet weak var textField: UITextField! {
+        didSet {
+            textField.delegate = self
+        }
+    }
 
     override func prepareForReuse() {
         super.prepareForReuse()
 
         textField.delegate = nil
-        unitLabel.text = nil
+        unitLabel?.text = nil
+    }
+    
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: true)
+        
+        if selected {
+            if textField.isFirstResponder {
+                textField.resignFirstResponder()
+            } else {
+                textField.becomeFirstResponder()
+            }
+        }
+    }
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        
+        contentView.layoutMargins.left = separatorInset.left
+        contentView.layoutMargins.right = separatorInset.left
+    }
+    
+    weak var delegate: TextFieldTableViewCellDelegate?
+    
+    // MARK: - UITextFieldDelegate
+    
+    func textFieldDidBeginEditing(_ textField: UITextField) {
+        delegate?.textFieldTableViewCellDidBeginEditing(self)
+    }
+    
+    func textFieldDidEndEditing(_ textField: UITextField) {
+        delegate?.textFieldTableViewCellDidEndEditing(self)
     }
 }
